Run RuboCop only once per CI run
Right now, RuboCop is called for every job in the CI matrix, however
there should never be a different result, depending on Ruby or Rails
version.

So we only run RuboCop once, in a Lint stage.

The `script` part was removed as the default is `bundle exec rake` which
is the same as `bundle exec rake test` in the case of this gem.