Restore controller action hash separator
The removal of the separator was introduced in 9e7744087.

I believe this to be a regression, as do others in the Pull Request
comments [1].

Given the commit message, it doesn't appear that the author appreciated
the reasons behind the double hash.

Other specs assert that the controller and action name is separated by a
hash, not a space:

    $ grep -r "home#index" test
    test/exception_notifier/email_notifier_test.rb:    assert_equal '[Dummy ERROR] home#index (ZeroDivisionError) "divided by 0"', @mail.subject
    test/exception_notifier/email_notifier_test.rb:      A ZeroDivisionError occurred in home#index:
    test/exception_notifier/modules/formatter_test.rb:    assert_equal 'A *RuntimeError* occurred in *home#index*.', formatter.subtitle
    test/exception_notifier/modules/formatter_test.rb:    assert_equal 'home#index', formatter.controller_and_action

    $ grep -Er "#\{kontroller.controller_name\}##\{kontroller.action_name\}" test
    test/exception_notifier/slack_notifier_test.rb:      text += " *was processed by* `#{kontroller.controller_name}##{kontroller.action_name}`" if kontroller

The test for EmailNotifier was introduced nearly 2 years later in
eb173627, so the regression was not caught at the time. The test author
mirrored the incorrect behaviour, presumably to get some improved level
of coverage based on the existing code.

[1] https://github.com/smartinez87/exception_notification/pull/398